requirements_dcgen.txt
selenium==4.18.1
pillow==10.2.0
tqdm==4.66.2
pandas==2.2.1
pillow-avif-plugin==1.4.3
matplotlib
openai==1.13.3
ipywidgets
ipykernel
scikit-image
google-generativeai
pyqt5
lxml==5.2.2
BeautifulSoup4 
requests
flask==3.0.3
rapidfuzz==3.6.2
anthropic
httpx==0.27.2 